cext: fix Windows warning about implicit conversion of 32-bit shift to 64 bit
    mercurial/cext/revlog.c(1541) : warning C4334: '<<' : result of 32-bit shift
        implicitly converted to 64 bits (was 64-bit shift int ended?)

+++ mercurial/cext/revlog.c
@@ -1535,7 +1535,7 @@ static PyObject *find_deepest(indexObject *self, PyObject *revs)
 		goto bail;
 	}
 
-	interesting = calloc(sizeof(*interesting), 1 << revcount);
+	interesting = calloc(sizeof(*interesting), ((size_t)1) << revcount);
 	if (interesting == NULL) {
 		PyErr_NoMemory();
 		goto bail;
